pyproject.toml
[tool.poetry]
name = "python-blame"
version = "0.1.2"
description = "A Python tool and a library for blaming Python code"
authors = ["Roman Inflianskas <infroma@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.scripts]
python-blame = "python_blame:main"

[tool.poetry.dependencies]
python = "^3.5"
plumbum = "^1.6"
astroid = "^2.2"
"ruamel.yaml" = "^0.15.89"
docopt = "^0.6.2"
poetry-version = "^0.1.3"

[tool.poetry.dev-dependencies]
ipython = "^7.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
